--- src/main/java/woowacourse/auth/dto/TokenResponse.java
@@ -1,16 +1,0 @@
-package woowacourse.auth.dto;
-
-public class TokenResponse {
-    private String token;
-
-    public TokenResponse() {
-    }
-
-    public TokenResponse(String token) {
-        this.token = token;
-    }
-
-    public String getToken() {
-        return token;
-    }
-}
