+++ src/main/java/woowacourse/auth/ui/AuthController.java
@@ -3,7 +3,6 @@ package woowacourse.auth.ui;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 import woowacourse.auth.application.AuthService;
 import woowacourse.auth.dto.SignInRequest;
