+++ src/main/java/woowacourse/shoppingcart/dto/CustomerResponse.java
@@ -1,4 +1,4 @@
-package woowacourse.auth.dto;
+package woowacourse.shoppingcart.dto;
 
 public class CustomerResponse {
